Remove RecapStuffListView and reuse StuffListView
